+++ tests/test_hf_upload.py
@@ -1,4 +1,4 @@
-import hf_upload
+from utils import hf_upload
 from datetime import datetime
